Internationalization of Hero and About sections
All the texts have been translated.
Correct display of the Project info is a WIP
