+++ messages/en.json
@@ -1,0 +1,26 @@
+{
+  "Hero": {
+    "subtitle": "Ingeniera Multimedia Multimedia Engineer"
+  },
+  "About": {
+    "about0": "Hi! I am Irene, a Spanish multimedia engineering graduate with more than 3 years of work experience as a Software Engineer. Recently moved to Shinjuku, I’m looking for new opportunities in Tokyo!",
+    "about1": "My passion lies in the impact of digital media on our daily lives and the diverse creative ways in which people can interact with them.",
+    "about2": "Aside from programming, in my free time I enjoy playing video games curled up with my cat, fiber arts and learning about fashion history.",
+    "about3": "I have mainly worked with Java and Angular using Typescript, building the UX of aeronautical applications. However, I am always looking forward to acquiring new skills, currently looking forward to expand my Frontend tech stack.",
+    "download": "Download CV"
+  },
+  "Projects": {
+    "p0_title": "Personal Website",
+    "p1_title": "Veronte Suite",
+    "p2_title": "Bachelor’s Degree Thesis",
+    "p3_title": "Go! Go!! Robot Brawl",
+    "p0_description": "A place to showcase my current projects and knowledge. Hope you like it!",
+    "p1_description": "I had a major role in the development and maintenance of 12 Java-based aeronautic apps during a critical turning point in the company's transition to a new ecosystem. These apps, are used to visually configure the parameters required on each stage of a UAV autopilot calibration and flight monitoring.\n\nAdditionally, I worked on single-page web applications, for creating several UI cartographic tools to facilitate the design of the flight missions, and widgets to visualize and monitor the status of the platform.",
+    "p2_description": "As my Bachelor’s Degree Thesis, I created this immersive cultural experience made with Unreal Engine.\n\nThe environment is based on the City of Arts and Sciences complex located in Valencia, Spain. In it, the user can walk around L'Umbracle and admire the other astonishing architectural structures in the area while interacting with several points with information about the complex.",
+    "p3_description": "Go! Go!! Robot Brawl is a cartoon styled fighting game developed entirely from scratch using C++ and OpenGL with a small group as part of our Degree’s Final Group Project.\n\nMy main role was the creation of the graphics engine, shaders, and core gameplay mechanics as well as planning, documenting and reviewing the different sprints.\n\nFor our work, the University of Alicante granted us honors in both Post-production and Computer Graphics subjects.",
+    "p0_alt": "Website fluid background effect with decorative stars",
+    "p1_alt": "Veronte Ops world map and logo",
+    "p2_alt": "Close up of the textures and main structure of the Umbracle project",
+    "p3_alt": "The game's logo on top of a plain background"
+  }
+}
