+++ next.config.ts
@@ -8,5 +8,5 @@ const nextConfig: NextConfig = {
     unoptimized: true,
   },
 };
-
+module.exports = nextConfig;
 export default nextConfig;
