+++ package.json
@@ -12,6 +12,7 @@
     "@svgr/webpack": "^8.1.0",
     "clsx": "^2.1.1",
     "next": "15.1.2",
+    "next-intl": "^3.26.4",
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
   },
