+++ src/app/[locale]/global.css
