+++ src/app/[locale]/providers.tsx
@@ -1,0 +1,21 @@
+"use client";
+
+import { NextIntlClientProvider } from "next-intl";
+
+function Providers({
+  children,
+  locale,
+  messages,
+}: {
+  children: React.ReactNode;
+  locale: string;
+  messages: Record<string, string>;
+}) {
+  return (
+    <NextIntlClientProvider locale={locale} messages={messages}>
+      {children}
+    </NextIntlClientProvider>
+  );
+}
+
+export default Providers;
