+++ src/app/static-info/math-utils.tsx
